rename package codermcp -> mcp
