+++ mcp/resource.go
@@ -1,4 +1,4 @@
-package codermcp
+package mcp
 
 import (
 	"context"
